[#46] 학교정보 수정작업 완료
